Remove `_plugins` and `manage_plugins.py`
