--- src/skimage/io/_plugins/fits_plugin.ini
@@ -1,3 +1,0 @@
-[fits]
-description = FITS image reading via PyFITS
-provides = imread, imread_collection
