--- src/skimage/io/_plugins/gdal_plugin.ini
@@ -1,3 +1,0 @@
-[gdal]
-description = Image reading via the GDAL Library (www.gdal.org)
-provides = imread
