--- src/skimage/io/_plugins/imageio_plugin.ini
@@ -1,3 +1,0 @@
-[imageio]
-description = Image reading via the ImageIO Library
-provides = imread, imsave
