--- src/skimage/io/_plugins/imread_plugin.ini
@@ -1,3 +1,0 @@
-[imread]
-description = Image reading and writing via imread
-provides = imread, imsave
