--- src/skimage/io/_plugins/matplotlib_plugin.ini
@@ -1,3 +1,0 @@
-[matplotlib]
-description = Display or save images using Matplotlib
-provides = imshow, imread, imshow_collection, _app_show
