--- src/skimage/io/_plugins/pil_plugin.ini
@@ -1,3 +1,0 @@
-[pil]
-description = Image reading via the Python Imaging Library
-provides = imread, imsave
