--- src/skimage/io/_plugins/simpleitk_plugin.ini
@@ -1,3 +1,0 @@
-[simpleitk]
-description = Image reading and writing via SimpleITK
-provides = imread, imsave
