--- src/skimage/io/_plugins/tifffile_plugin.ini
@@ -1,3 +1,0 @@
-[tifffile]
-description = Load and save TIFF and TIFF-based images using tifffile.py
-provides = imread, imsave
