+++ src/skimage/io/meson.build
@@ -3,7 +3,6 @@ python_sources = [
   '_image_stack.py',
   '_io.py',
   'collection.py',
-  'manage_plugins.py',
   'sift.py',
   'util.py'
 ]
@@ -13,5 +12,3 @@ py3.install_sources(
   pure: false,             # Will be installed next to binaries
   subdir: 'skimage/io'  # Folder relative to site-packages to install to
 )
-
-subdir('_plugins')
